requirements.txt
google-genai>=1.0.0
sendgrid>=6.11.0
supabase>=2.6.0
pillow>=10.3.0
pypdf>=4.2.0
pdf2image>=1.17.0
python-dotenv>=1.0.1
streamlit